requirements.txt
# python version: 3.12.3
beautifulsoup4==4.12.3
contractions==0.1.73
Flask==3.1.0
flask_cors==5.0.1
Flask_MySQLdb==2.0.0
googletrans_py==4.0.0
joblib==1.4.2
keras==3.6.0
llama_index==0.12.22
nest_asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
Sastrawi==1.0.1
scikit_learn==1.5.2
simpletransformers==0.70.1
swifter==1.4.0
tensorflow==2.17.0
yfinance==0.2.54
torch==2.4.1
ipython==7.23.1
loguru==0.7.3
pytest==8.4.1
pytest-flask==1.3.0
SQLAlchemy==2.0.43

